ECOMM-112 refactor: separate folders Profile files
